Refactoring Gameta

+++ gameta/repos.py
@@ -6,7 +6,8 @@ from typing import Dict, Optional
 import click
 from git import Repo, GitError
 
-from . import gameta_cli, gameta_context, GametaContext
+from .cli import gameta_cli
+from .context import gameta_context, GametaContext
 
 
 __all__ = ['repo_cli']
@@ -84,6 +85,7 @@ def add(context: GametaContext, name: str, url: str, path: str, overwrite: bool)
         context.repositories[name] = {
             'url': url,
             'path': path,
+            '__metarepo__': False
         }
         context.add_gitignore(path)
         context.export()
@@ -120,13 +122,20 @@ def delete(context: GametaContext, name: str, clear: bool) -> None:
             click.echo(f"Repository {name} does not exist in the .meta file, ignoring")
             return
 
+        # Prevent user from deleting primary metarepo
+        if context.is_primary_metarepo(name):
+            raise click.ClickException(f"Cannot delete repository {name} as it is a metarepo")
+
         if clear:
             click.echo(f"Clearing repository {name} locally")
             rmtree(join(context.project_dir, context.repositories[name]["path"]), ignore_errors=True)
+
         context.remove_gitignore(context.repositories[name]["path"])
         del context.repositories[name]
         context.export()
         click.echo(f"Repository {name} successfully deleted")
+    except click.ClickException:
+        raise
     except Exception as e:
         raise click.ClickException(f"{e.__class__.__name__}.{str(e)}")
